Refactor migrations to align with enterprise repo
changelog: none
