+++ store/mongo/datastore_mongo.go
@@ -20,9 +20,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/mendersoftware/go-lib-micro/identity"
-	"github.com/mendersoftware/go-lib-micro/log"
-	"github.com/mendersoftware/go-lib-micro/mongo/migrate"
 	"github.com/mendersoftware/go-lib-micro/mongo/uuid"
 	mstore "github.com/mendersoftware/go-lib-micro/store"
 	"github.com/pkg/errors"
@@ -38,8 +35,6 @@ import (
 )
 
 const (
-	DbVersion      = "1.0.0"
-	DbName         = "useradm"
 	DbUsersColl    = "users"
 	DbTokensColl   = "tokens"
 	DbSettingsColl = "settings"
@@ -302,74 +297,6 @@ func (db *DataStoreMongo) SaveToken(ctx context.Context, token *jwt.Token) error
 
 	if err != nil {
 		return errors.Wrap(err, "failed to store token")
-	}
-
-	return nil
-}
-
-// MigrateTenant migrates a single tenant database.
-func (db *DataStoreMongo) MigrateTenant(ctx context.Context, version string, tenant string) error {
-	ver, err := migrate.NewVersion(version)
-	if err != nil {
-		return errors.Wrap(err, "failed to parse service version")
-	}
-
-	tenantCtx := identity.WithContext(ctx, &identity.Identity{
-		Tenant: tenant,
-	})
-
-	m := migrate.SimpleMigrator{
-		Client:      db.client,
-		Db:          mstore.DbFromContext(tenantCtx, DbName),
-		Automigrate: db.automigrate,
-	}
-	migrations := []migrate.Migration{
-		&migration_1_0_0{
-			ds:  db,
-			ctx: ctx,
-		},
-	}
-
-	err = m.Apply(tenantCtx, *ver, migrations)
-	if err != nil {
-		return errors.Wrap(err, "failed to apply migrations")
-	}
-	return nil
-}
-
-func (db *DataStoreMongo) Migrate(ctx context.Context, version string) error {
-	l := log.FromContext(ctx)
-
-	dbs := []string{DbName}
-
-	if db.multitenant {
-		l.Infof("running migrations in multitenant mode")
-
-		tdbs, err := migrate.GetTenantDbs(ctx, db.client, mstore.IsTenantDb(DbName))
-		if err != nil {
-			return errors.Wrap(err, "failed go retrieve tenant DBs")
-		}
-		dbs = tdbs
-	} else {
-		l.Infof("running migrations in single tenant mode")
-	}
-
-	if db.automigrate {
-		l.Infof("automigrate is ON, will apply migrations")
-	} else {
-		l.Infof("automigrate is OFF, will check db version compatibility")
-	}
-
-	for _, d := range dbs {
-		l.Infof("migrating %s", d)
-
-		// if not in multi tenant, then tenant will be "" and identity
-		// will be the same as default
-		tenant := mstore.TenantFromDbName(d, DbName)
-
-		if err := db.MigrateTenant(ctx, version, tenant); err != nil {
-			return err
-		}
 	}
 
 	return nil
